DRTII-1346 Tidy & add missing max pax where needed
