+++ src/main/resources/application.conf
@@ -9,6 +9,8 @@ include "jdbc.conf"
 
 options {
   job-name: ${?JOB_NAME}
+  dry-run: false
+  dry-run: ${?DRY_RUN}
   training {
     days-of-data: 150
     days-of-data: ${?TRAINING_DAYS_OF_DATA}
@@ -22,12 +24,9 @@ options {
 spark.testing.memory = 400000000
 
 aws {
-  access-key-id = "some-access-key-id"
   access-key-id = ${?AWS_ACCESS_KEY_ID}
-  secret-access-key = "some-secret-access-key"
   secret-access-key = ${?AWS_SECRET_ACCESS_KEY}
   s3 {
-    bucket = "some-bucket"
     bucket = ${?AWS_S3_BUCKET}
   }
 }
