+++ src/main/resources/log4j2.properties
@@ -4,8 +4,8 @@ rootLogger=INFO, STDOUT
 # Define console appender
 appender.console.name = STDOUT
 appender.console.type = Console
-appender.console.layout.type = PatternLayout
-appender.console.layout.pattern = [%-5level] %d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %c{1} - %msg%n
+appender.console.layout.type = JsonTemplateLayout
+appender.console.layout.eventTemplateUri = classpath:EcsLayout.json
 
 # Set log level for specific packages or classes
 # For example, to set the log level to WARN for the org.apache.spark package:
