+++ src/main/scala/uk/gov/homeoffice/drt/analytics/prediction/flights/FlightValueExtractionActor.scala
@@ -16,14 +16,13 @@ import uk.gov.homeoffice.drt.ports.Terminals.Terminal
 import uk.gov.homeoffice.drt.protobuf.messages.CrunchState.{FlightWithSplitsMessage, FlightsWithSplitsDiffMessage, FlightsWithSplitsMessage}
 import uk.gov.homeoffice.drt.protobuf.messages.FlightsMessage.UniqueArrivalMessage
 import uk.gov.homeoffice.drt.protobuf.serialisation.FlightMessageConversion.flightWithSplitsFromMessage
-import uk.gov.homeoffice.drt.time.{SDate, UtcDate}
+import uk.gov.homeoffice.drt.time.UtcDate
 
-import scala.collection.immutable
 import scala.concurrent.{ExecutionContextExecutor, Future}
 
 
 object FlightValueExtractionActor {
-  case class PreProcessingFinished(byArrivalKey: Map[ArrivalKey, Arrival])
+  case class PreProcessingFinished(arrivals: Iterable[Arrival])
 
   val noCtaOrDomestic: FlightWithSplitsMessage => Boolean = msg => !isDomesticOrCta(PortCode(msg.getFlight.getOrigin))
 }
@@ -43,7 +42,7 @@ object FlightMessageConversions {
 trait FlightValueExtractionActorLike {
   private val log = LoggerFactory.getLogger(getClass)
 
-  var byArrivalKey: Map[ArrivalKey, Arrival] = Map()
+  var byArrivalKey: Map[ArrivalKey, Arrival] = Map.empty[ArrivalKey, Arrival]
 
   val extractValues: Arrival => Option[(Double, Seq[String], Seq[Double], String)]
   val extractKey: Arrival => Option[WithId]
@@ -67,14 +66,15 @@ trait FlightValueExtractionActorLike {
     msg.removals.foreach(processRemovalMessage)
   }
 
-  def extractions(byArrivalKeyProcessed: Map[ArrivalKey, Arrival]): Map[WithId, immutable.Iterable[(Double, Seq[String], Seq[Double], String)]] =
+  def extractions(byArrivalKeyProcessed: Map[ArrivalKey, Arrival]): Map[WithId, Iterable[(Double, Seq[String], Seq[Double], String)]] =
     byArrivalKeyProcessed
       .groupBy {
-        case (_, msg) => extractKey(msg)
+        case (_, arrival) => extractKey(arrival)
       }
       .collect {
         case (Some(key), arrivals) =>
-          val examples = arrivals
+          val examples = arrivals.values
+            .map(a => (a.unique, a))
             .map { case (_, arrival) => extractValues(arrival) }
             .collect { case Some(value) => value }
           (key, examples)
@@ -85,7 +85,7 @@ class FlightValueExtractionActor(val terminal: Terminal,
                                  val date: UtcDate,
                                  val extractValues: Arrival => Option[(Double, Seq[String], Seq[Double], String)],
                                  val extractKey: Arrival => Option[WithId],
-                                 val preProcessing: (UtcDate, Map[ArrivalKey, Arrival]) => Future[Map[ArrivalKey, Arrival]],
+                                 val preProcessing: (UtcDate, Iterable[Arrival]) => Future[Iterable[Arrival]],
                                 ) extends TerminalDateActor[Arrival] with PersistentActor with FlightValueExtractionActorLike {
   private val log: Logger = LoggerFactory.getLogger(getClass)
 
@@ -93,15 +93,15 @@ class FlightValueExtractionActor(val terminal: Terminal,
 
   implicit val ec: ExecutionContextExecutor = context.dispatcher
 
-  var valuesWithFeaturesByExtractedKey: Map[WithId, Iterable[(Double, Seq[String], Seq[Double], String)]] = Map()
-  var preProcessingFinished = false
-  var stateRequests: List[ActorRef] = List()
+  private var valuesWithFeaturesByExtractedKey: Map[WithId, Iterable[(Double, Seq[String], Seq[Double], String)]] = Map()
+  private var preProcessingFinished = false
+  private var stateRequests: List[ActorRef] = List()
 
   override def receiveRecover: Receive = {
     case SnapshotOffer(_, ss) => processSnapshot(ss)
 
     case RecoveryCompleted =>
-      preProcessing(date, byArrivalKey)
+      preProcessing(date, byArrivalKey.values)
         .map(self ! PreProcessingFinished(_))
         .recover(e => log.error(s"Failed to pre-process flights for $persistenceId", e))
 
@@ -110,8 +110,8 @@ class FlightValueExtractionActor(val terminal: Terminal,
   }
 
   override def receiveCommand: Receive = {
-    case PreProcessingFinished(byArrivalKeyProcessed) =>
-      valuesWithFeaturesByExtractedKey = extractions(byArrivalKeyProcessed)
+    case PreProcessingFinished(arrivals) =>
+      valuesWithFeaturesByExtractedKey = extractions(arrivals.map(a => (ArrivalKey(a), a)).toMap)
       if (stateRequests.nonEmpty) {
         stateRequests.foreach(_ ! valuesWithFeaturesByExtractedKey)
         stateRequests = List()
