+++ src/main/scala/uk/gov/homeoffice/drt/analytics/services/ArrivalsHelper.scala
@@ -4,7 +4,6 @@ import akka.actor.ActorSystem
 import akka.pattern.ask
 import akka.util.Timeout
 import org.slf4j.LoggerFactory
-import uk.gov.homeoffice.drt.actor.TerminalDateActor.ArrivalKey
 import uk.gov.homeoffice.drt.analytics.Arrivals
 import uk.gov.homeoffice.drt.analytics.actors.{ArrivalsActor, FeedPersistenceIds, GetArrivals}
 import uk.gov.homeoffice.drt.arrivals.Arrival
@@ -15,28 +14,35 @@ import scala.concurrent.{ExecutionContext, Future}
 object ArrivalsHelper {
   private val log = LoggerFactory.getLogger(getClass)
 
-  val noopPreProcess: (UtcDate, Map[ArrivalKey, Arrival]) => Future[Map[ArrivalKey, Arrival]] = (_, a) => Future.successful(a)
+  val noopPreProcess: (UtcDate, Iterable[Arrival]) => Future[Iterable[Arrival]] = (_, a) => Future.successful(a)
 
   def populateMaxPax()(implicit
                        system: ActorSystem,
                        ec: ExecutionContext,
-                       timeout: Timeout): (UtcDate, Map[ArrivalKey, Arrival]) => Future[Map[ArrivalKey, Arrival]] =
+                       timeout: Timeout): (UtcDate, Iterable[Arrival]) => Future[Iterable[Arrival]] =
     (date, arrivals) => {
-      val pctWithMaxPax = (100 * arrivals.values.count(_.MaxPax.isDefined).toDouble / arrivals.size).round
-      if (pctWithMaxPax < 80) {
-        log.debug(s"Only $pctWithMaxPax% of arrivals have max pax for $date, populating")
+      def pctWithMaxPax(arrivals: Iterable[Arrival]): Int = (100 * arrivals.count(_.MaxPax.isDefined).toDouble / arrivals.size).round.toInt
+
+      val pctOk = pctWithMaxPax(arrivals)
+      if (pctOk < 80) {
+        log.info(s"Only $pctOk% of arrivals have max pax for $date, populating")
         val arrivalsActor = system.actorOf(ArrivalsActor.props(FeedPersistenceIds.forecastBase, SDate(date)))
         arrivalsActor
           .ask(GetArrivals(SDate(date), SDate(date).addDays(1))).mapTo[Arrivals]
           .map { baseArrivals =>
             arrivalsActor ! akka.actor.PoisonPill
 
-            arrivals.view.mapValues {
+            arrivals.map {
               arr =>
                 val maybeArrival = baseArrivals.arrivals.get(arr.unique)
                 val maybeMaxPax = maybeArrival.flatMap(_.maxPax)
                 arr.copy(MaxPax = maybeMaxPax)
-            }.toMap
+            }
+          }
+          .map{ a =>
+            val pctOk = pctWithMaxPax(a)
+            log.info(s"Populated max pax for $date, now $pctOk% have max pax")
+            a
           }
           .recover {
             case t =>
